java launcher: move to dlopen() based interface
Our launcher is no longer linked to libjvm.so now.  Instead, it searches
the library in a path relative to the executable, and dynamically binds
to it.

The advantage is that you no longer have to play tricks with
LD_LIBRARY_PATH, indeed, it should even work when you are not allowed to
fiddle with libraries.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>