Add plugins infrastructure
Most plugins we get probably need to be adjusted.  We plan to have a
consistent menu layout, no matter who wrote the plugin and what its
original plugins.config looked like.

The idea is that you put the .jar file into staged-plugins/, create
or modify a .config file there, too.  A simple "make" will put the
.config file as "plugins.config" into the .jar and move it into
plugins/.  Both the resulting .jar and the .config will be git-added.

Example:

	$ cp Primes_.jar staged-plugins/
	$ echo 'File, "Show Prime Numbers", Primes("")' \
		> staged-plugins/Primes_.config
	$ make

The result will be plugins/Primes_.jar, and both this file and
staged-plugins/Primes_.config will be automatically committed to the
git repository.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>