Add a plugin to get the class versions
Since we have two different JDK versions in Fiji, Java5 and Java6, we
have to be careful which class versions we ship.  The plugin
'Get_Class_Versions' should help.

Incidentally, this adds an infrastructure to have plugin sources in Fiji,
but outside of the plugins/ directory (which we will eventually ship).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>