RefreshScripts now looks through all of plugins/ for scripts
Previously you would specify a particular Plugins submenu with a flat
list of scripts, but we don't really want to have to have JRuby and
Jython scripts in special subdirectories - they should just appear
like any other command whereever is most appropriate.

Note that although the part that ensures that a given submenu
exists seems to work OK, I should replace it with the equivalent
function from ImageJA to avoid code duplication.  It's probably OK
for the moment, anyway.