Add a submodule for MicroManager1.1
MicroManager is everything, except easy to compile from scratch.  So it
makes tons of sense to try to make it as painless as possible to compile
it within Fiji.

This adds a "mm" target to the Makefile to compile everything.  It relies
on the "fiji" branch of our version of MicroManager, which is the git-svn
mirrored "trunk" of the upstream, with a few patches to make it compile
within Fiji.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>