Make the ClassVersion check more useful
This commit converts the class version checker into a non-plugin which
gets called via the call() function of ImageJ's macro language.

Also, it aggregates the number of classes by class version, prints the
Java version which is required by that class version, and it outputs
only those using a too-new Java version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>