Move precompiled binaries into the directory precompiled/
It is not convenient to have binaries for all supported platforms in
the root directory: tab-completion would match all of them.

Therefore, move them into the directory precompiled/.

Storing the precompiled binaries is only necessary until we can
cross-compile them (which is not helped by the fact that the odcctools
do not compile on x86_64... D'oh!).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>