Fiji launcher on MacOSX: work around Tiger 64-bit issues
On Tiger, next to no library is 64-bit.  Least of all, CoreFoundation
and JavaVM.  So the universal binary would fail on load.  Additionally,
there is no option to force a certain cpu_type.  There is one for
Leopard, but then, Leopard does not need it, since it has all the
64-bit libraries we need!

So work around that by having a minimal launcher which finds out the
OS version, and a second universal binary which only includes 32-bit
versions.

Greetings to Cupertino.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>