Fake: include all .class files for a set of .java files in the .jar
Now we inspect all .class files for the given .java files, and
iterate over _all_ referenced classes.  If the corresponding
.class files are found in the same location as the .java file,
we include it in the .jar file.

This change is needed to support .java files which contain (package
local) classes with _another_ name.  I.e. if you have

	class XYZ {
		...
	}

in a file named ABC.java, the class "XYZ" will be created, and with
this commit, it will be included in the .jar file, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>