Fake.sh: work around Windows being unable to overwrite in-use files
When compiling fiji.exe or fake.jar with "fiji --fake", Windows punts
because it is not able to overwrite/delete/rename files that are in
use (at least with Java, since it memory-maps the files).

To work around that, teach Fake.sh to fall back to precompiled/ to
compile fiji or fake.jar.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>