Fake: fix classPath handling on Windows
As with so many things, path separators are different on Windows.  So
we have to handle them specially.  However, in the Fakefile we do _not_
want that Windows crap, so we use ":" as path separators there.  Not
";", which never wanted to be anything else than the end of a command.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>