Fiji launcher (cygwin): link with -mno-cygwin
Otherwise, cygwin1.dll will be linked.

While at it, make precompiled/fiji-win32.exe executable again

MinGW does not complain, since it has a peculiar way to handle the
executable bit: if the file ends in .exe, .cmd, .bat, or starts with
"#!", it is executable.  You just cannot change that.

Cygwin, however, _wants_ the exexcutable bit set, otherwise it refuses
to execute the program, even if it has a .exe extension.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>