Move a lot of plugins.configs from src-plugins/ to staged-plugins/
The plugins.config files are meant to live in staged-plugins/<name>.config
so that there is a central place to look what is where.

Note that there is a valid reason to have plugins.config in src-plugins/:
ImageJ does not yet allow as flexible menu overriding as we do in Fiji.

So if the plugin is meant to be distributed externally, you can add the
"external" plugins.config directly in src-plugins/, providing a .config
in staged-plugins/ that will override the "external" one in Fiji.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>