Fake: allow overriding variables from the command line
Now, "./fiji --fake verbose=true [<target>]" will override the setting
in the Fakefile and set the variable "verbose" to "true".  Note that
this does not override the specializations, i.e. if the Fakefile has
a line

	VERBOSE(blub.jar)=false

you need to use "./fiji --fake 'verbose(blub.jar)=true' blub.jar" to
override the setting.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>