Fake: always call javac with all .java files
We had some cute hack in Fake to call javac only with the .java files
that were newer than the target .jar file.

However, this broke down when some classes in src-plugins/ needed other
(up-to-date) classes in src-plugins/, since the classpath is not set to
src-plugins/.

Always compiling all the specified .java files works, though, so do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>