Fakefile: add targets for all applications, and packages
In addition to the targets we already supported, there is an 'app-all',
which includes the JREs and launchers for all platforms, and an
'app-nojre', which includes the launchers for all platforms, but
without JREs.

As a further benefit, it does not need the 'zip' executable anymore
when run from Jython, using Java's own Zip implementation.

In short: the following targets are supported now:

	app-linux, app-linux64, app-macosx app-win32, app-win64,
	app-all, app-nojre

	fiji-macosx.dmg,
	fiji-win32.zip, fiji-win64.zip, fiji-linux.zip, fiji-linux64.zip,
	fiji-all.zip, fiji-nojre.zip,
	fiji-linux.tar.bz2, fiji-linux64.tar.bz2, fiji-all.tar.bz2,
	fiji-nojre.tar.bz2

As a convenience, the 'dmg', 'tar' and 'zip' targets build the
appropriate package for the host platform (if available).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>