Re-Fix icon on MacOSX
In a5ae68c5(Fakefile: add targets for all applications, and packages),
we changed the way fiji_dir is discovered in a MacOSX application, to
avoid symlinks (so that the app can actually be put on a FAT-formatted
USB stick).

Unfortunately, we forgot to update the icon discovery, which is fixed
by this commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>