Fiji launcher: open a console on Windows on demand
On Windows, there is this pretty braindead convention that a program is
either a console program or a non-console program.  However, we often
need to output some error messages to stderr, but do not want a console
if no error occurred.

So open a console on demand (on XP and later even trying to use the parent
process' console) when stderr was written to.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>