Add three Jython examples
This commit adds three examples how you can develop early versions of
a plugin in a quick and dirty way.

Python is not strictly-typed, so it is convenient to hack your way
through a Java API.

Also, if you use Jython, you do not need to catch exceptions, as they
will just abort the whole program.

Later on, however, you should turn this into a strictly-typed, and
safe-to-use plugin (for example, it is not nice to the user to get an
exception window...)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>