Move tracked .jar files from plugins/ into staged-plugins/
We ship with a few third party plugins, but we optionally insert custom
plugins.config files.  So the .jar files that are in plugins/ are actually
generated, so we should commit the original .jar files in staged-plugins/
instead.  So do this.

This is a late companion commit to 0470894(Untrack generated files).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>