Fake: optionally export an Eclipse project
Eclipse seems to be the Malbolge of IDEs.  They make it actively hard
to ship any general project settings with your project, except if you
created the project in Eclipse to begin with, and even then, chances are
that you have an absolute path in those settings by mistake.

In other words, Eclipse is a major pain in the arse.

So this attempt to support generating an easy-to-ship, easy-to-install
Eclipse environment should make things much easier for us poor Fiji
people.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>