Fake: force usage of jars/javac.jar for compiling
When running on older Java, we still want to use our own javac to be
able to build classes for newer Java.  So force loading _our_ javac.

This is achieved by extending the JarClassLoader to a saner interface
than was done when we changed Fake to fall back to precompiled Fake in
case that was more up-to-date than the current one.

It is problematic since most classes that require Java 1.5 or later do
not only rely on newer language features, but also newer API features.
While Retrotranslator can help running those, it does not give javac
the full classes it would need to resolve the new features (e.g. the
StringBuilder).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>