Added the plugin "Fiji Updater"
The Fiji Updater lives in the Help menu, and can be used to check for
updates to Fiji components.

Since .jar files are not kept open by Sun's Java virtual machine, the
.jar files cannot be updated from within ImageJ.  Therefore, all the
files that need updating are stored in the update/ directory.

The Fiji launcher is an exception: it is updated directly (after renaming
the existing launcher, which is needed on Windows).

It is the Fiji launcher's responsibility is to move the contents of the
update/ directory to their intended place upon startup.  Thus, an update
requires a restart of Fiji.

The updater does not yet cope with JDK.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>