Fiji launcher: use ImageJ port "7" by default
Many users try to run Fiji while an existing ImageJ instance is running.
Then they are confused that Fiji exits right away.

This also confused the author of this patch, wasting a whole hour on the
issue.

Use a different port for Fiji by default, so that existing ImageJ
instances do not confuse Fiji (and users and developers).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>