View5D: move it into Plugins>Analyze, and into the java package "view5d"
The plugin View5D was not in a java package, and worse, it used pretty
reserved class names, such as "ImageCanvas".  This made something like

	./fiji --javac Dynamic_Profiler.java

fail, since the Dynamic_Profiler wanted to access ImageJ's ImageCanvas.

So move it out of the way, and while at it, move the menu entry into
the "Analyze" menu.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>