Fiji launcher: be cleverer when discovering bundled java
If the compiled-in path to our preferred JDK does not exist, we now
just use the newest available.

To this end, we look into the newest subdirectory of java/ (which is
supposed to have the platform name), and then we look for the newest
subdirectory thereof (which is supposed to be the JDK directory).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>