Fake: support the .iso target fully
The .iso file target necessarily has to call an external program,
but the prerequisite listed is not the actual prerequisite, as we
want to tell Fake _which_ platform to make Fiji.app/ for.

This will happen quite often with targets calling a program, I guess,
so teach Fake that such targets should always run that program when
none of the prerequisites are actually existing files, by pretending
that it is always out-of-date.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>