Add an initial version of a LiveCD creator script
This script is able to generate a live CD on any recent Ubuntu/Intel
system.  It does so by creating a chroot'ed environment, and using the
fantastic bootcd package.  A few things are still to do:

	- generate /etc/passwd and friends, instead of reusing the
	  current systems' (spilling secrets),
	- use a different user than the current one,
	- install X11, Java, and an appropriate window manager,
	- copy Fiji over,
	- configure it so the desktop starts without logging in, and
	  start Fiji right away.

Also: this works only with i386 so far, need to investigate how to
support 64-bit (x86_64, _not_ ia64).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>