Fiji launcher (win32): "fix" --system
Unfortunately, java.exe is not able to reuse the console, even if it
purports to do so.  This is most likely a side effect of us attaching
to the console of the parent process, and Windows not properly recording
the fact.

So '--system' worked, except that it failed to provide any console output
to tell the user that it worked.

Since we cannot fix Windows, or for that matter, the Windows -mconsole vs
-mwindows insanity, we just release the console before launching system
java no matter what.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>