Fiji launcher: Try to work around IPv6 permission bug
It seems that in some setups, notably some OpenSUSE setups, IPv6 is
not really configured properly.  Unfortunately, this leads to a
three-minute delay in the startup of ImageJ, because it wants to
connect to the IPv6 loopback device, timing out because the connection
is dropped by the firewall(!).

So test if we can create an IPv6 socket and make a connect call without
a timeout, and set the Java property java.net.preferIPv4Stack to
true if we cannot.

Noticed by Ignacio Arganda-Carreras.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>