Add the submodule "jna"
JNA is a way to call functions in a shared library without the need to
use JNI directly (i.e. without the need for a C compiler).

Of course, the price to be paid for the relative ease compared to JNI
is that JNA is about 10-30 times slower.

Use it like this:

	import com.sun.jna.Library;
	import com.sun.jna.Native;

	public class Example {

	    public interface MyLib extends Library {
		void thisFunctionExistsInTheLibrary(int number, String string);
	    }

	    public static void main(String[] args) {
		MyLib myLib =
			(MyLib)Native.loadLibrary("my_lib.so", MyLib.class);

		myLib.thisFunctionExistsInTheLibrary(1, "one");
	    }
	}

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Conflicts:

	.gitignore
	.gitmodules
	Fakefile