Fiji build: extend up-to-date check to the classpath
This changes fixes the issue when two submodules depend on each other,
but only one was rebuilt.

Example: when mpicbg was changed, but TrakEM2 not, TrakEM2 would not
have been rebuilt because TrakEM2 does not have a target for
mpicbg, and therefore did not realize it has to be rebuilt, too.

Noticed both by Stephan Saalfeld and Albert Cardona.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>