Fiji launcher: consolidate the code to get the JAVA_HOME
We can override Fiji's idea of where java is by setting the environment
variable JAVA_HOME.  This can be overridden with the command line
option '--java-home=<location>'.

Instead of duplicating the code everywhere, introduce the functions
get_java_home() and get_jre_home().

This also fixes a bug where --javap would not find the tools.jar if
Fiji's default JDK was not found.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>