Add an example to help finding the dimensions of a raw image
If you have a raw image but do not know the real dimensions, you can use
"Examples>Find Dimension of Raw Image" to play with the width interactively.

Thanks to Jython's weak-typing, the code works for all image types.

At the same time, this script demonstrates how you can write an AWT Listener
in Jython: an adjustment listener written in plain Python reacts to every
change of the slider.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>