Used "git branch newbranch HEAD^^", re-deleted test.txt
Creates a new branch with 2nd previous commit as HEAD (Hence HEAD^^). May also specify commit tag if "HEAD^^^^" turns out to be a little too tedious?

NOTE: It means in this new branch that test.txt is not deleted, diary for 5th May is not fully updated.

Also NOTE: Here, test.txt is deleted again.