Refactoring stage 1: Use Layout Managers (BoxLayout)
Never knew it would be so time consuming. Components were divided up
and added to JPanels, which were each given BoxLayout. The result
is similar to previous one, though in the future I would improve on
it, perhaps using GridBagLayout. Top level JPanels are "Top" and
"Bottom" panel, which the former consists of "Left" and "Right"
panels as well. Also, TextPaneFormat code has been shifted over to
PluginManager itself.