Program now does dependency checking AND DISPLAYS any conflicts.
The algorithm has been tweaked a little, but remains largely the
same from previous commit in terms of logic. Displaying of info
is only through a System.out.println for now, to see if style of
output is acceptable before I move on to design a proper interface.
A lot of repeating code at both PluginManager and Controller
classes, will do a "DRY" rework in the near future.