PluginManager do the downloading, Updater do the uploading
After a loooonnng time of deliberation... It is decided that
PluginManager should not be in charge of uploading changes to
server. Updater shall do the updating, with Uploader class
becoming a "UI" class of sorts (No JFrame yet). Same remains for
PluginManager where its counterpart FrameManager is responsible for
UI display.