Re-add Help>Update Fiji
We advertized Help>Update Fiji widely to be the most convenient method
to keep Fiji up-to-date.  So we need to keep that plugin working.

In the future, Help>Update Fiji should call the Plugin Manager in a
special mode that automatically shows only the update-able files and
marks them for update.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>