Simplify the DependencyAnalyzer
The DependencyAnalyzer should not try to do too many things at the same
time.  It is fully sufficient that it analyzes the .jar files on which
the given .jar file depends.  No need to check if the plugin collection
knows about those .jar files.  And there is no need for that class to
know how the list will be stored eventually.

This patch also changes a few hardcoded List<PluginObject> instances
that I missed earlier.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>